fix frame in /nominate, 1st draft final flow
