+++ src/app/layout.tsx
@@ -9,32 +9,8 @@ import { Web3ModalProvider } from "@/app/_providers/web-3-modal-provider";
 import { config } from "@/config";
 import { Footer } from "@/shared/components/footer";
 import { Header } from "@/shared/components/header";
-import { FRAMES_BASE_PATH, appURL } from "@/shared/frames/utils";
 import { ThemeRegistry } from "@/shared/theme/theme-registry";
 import { AuthenticationProvider } from "./_providers/authentication-provider";
-import type { Metadata } from "next";
-
-const description = [
-  "BUILD is a meme and a social game designed to reward builders via",
-  "onchain nominations.",
-].join(" ");
-
-export async function generateMetadata(): Promise<Metadata> {
-  return {
-    title: "BUILD",
-    description: description,
-    openGraph: {
-      title: "BUILD",
-      description: description,
-      type: "website",
-      url: "https://build.top",
-      images: ["https://build.top/images/BUILD-thumbnail.jpg"],
-    },
-    other: {
-      ...(await fetchMetadata(new URL(FRAMES_BASE_PATH, appURL()))),
-    },
-  };
-}
 
 export default function RootLayout({
   children,
